Extract Generate.generateElmJson
