+++ lib/elm-test.js
@@ -78,9 +78,11 @@ function resolveFilePath(filename) {
   });
 }
 
-var pathToElmBinary = "elm";
+let pathToElmBinary;
 
-if (args.compiler !== undefined) {
+if (args.compiler === undefined) {
+  pathToElmBinary = "elm";
+} else {
   pathToElmBinary = path.resolve(args.compiler);
 
   if (!pathToElmBinary) {
@@ -154,12 +156,8 @@ else if (args._[0] == "init") {
 
 var projectRootDir;
 var testRootDir;
-var elmJsonPath;
-var generatedCodeDir = path.resolve(
-  path.join("elm-stuff", "generated-code", "elm-community", "elm-test")
-);
 
-function runTests(testFile) {
+function runTests(generatedCodeDir/*:string*/, testFile/*:string*/) {
   const dest = path.resolve(path.join(generatedCodeDir, "elmTestOutput.js"));
 
   return Compile.compile(
@@ -255,18 +253,10 @@ function runElmTest() {
     process.exit(1);
   }
 
-  projectRootDir = path.resolve(Runner.findNearestElmPackageDir(initialTestFilePaths));
-  testRootDir = path.resolve(path.join(projectRootDir, "tests"));
-  generatedCodeDir = path.resolve(
-    path.join(projectRootDir, "elm-stuff", "generated-code", "elm-explorations", "test")
-  );
-  elmJsonPath = path.resolve(path.join(projectRootDir, "elm.json"));
-
-  // TOOD: error if "<project>/tests" does not exist
-
-  var returnValues = generatePackageJson(filePathArgs);
-  var generatedSrc = returnValues[0];
-  var sourceDirs = returnValues[1];
+  var returnValues = Generate.generateElmJson(pathToElmBinary, filePathArgs);
+  const generatedCodeDir = returnValues[0];
+  const generatedSrc = returnValues[1];
+  const sourceDirs = returnValues[2];
 
   function run() {
     var testFilePaths = _.flatMap(globs, resolveFilePath);
@@ -286,6 +276,7 @@ function runElmTest() {
           return generateAndRunTests(
             runnableTests,
             filePathArgs,
+            generatedCodeDir,
             generatedSrc,
             getGlobs
           );
@@ -342,181 +333,8 @@ function runElmTest() {
   }
 }
 
-function generatePackageJson(filePathArgs) {
-  // TODO we don't want to do this every single time. Instead,
-  // verify that the generated elm.json is there, with the
-  // expected version number. Iff the version number is wrong, regenerate.
-  var generatedSrc = path.join(generatedCodeDir, "src");
-  fs.mkdirpSync(generatedCodeDir);
-  fs.mkdirpSync(generatedSrc);
 
-  var projectElmJson = {};
-  try {
-    projectElmJson = fs.readJsonSync(elmJsonPath);
-  } catch (err) {
-    console.error("Error reading elm.json: " + err);
-    process.exit(1);
-  }
-  var isPackageProject = projectElmJson.type === "package";
-
-  var testElmJson = {
-    type: "application",
-    "source-directories": [], // these are added below
-    "elm-version": "0.19.0",
-    "dependencies": {
-      "direct": {},
-      "indirect": {}
-    },
-    "test-dependencies": {
-      "direct": {},
-      "indirect": {}
-    }
-  };
-
-  function addDirectDependencies(deps) {
-    Object.keys(deps).forEach(function(name) {
-      var version = deps[name];
-      if (isPackageProject) {
-        // Use the lowest version in the range.
-        // NOTE: technically this doesn't work if someone does something weird like:
-        //
-        // "2.0.0 < v < 3.0.0"
-        //
-        // ...but we're choosing not to support that right now.
-        version = version.split(" ")[0];
-      }
-      if (testElmJson["dependencies"]["direct"].hasOwnProperty(name)) {
-        var existingVersion = testElmJson["dependencies"]["direct"][name];
-
-        // If we have a clash, choose the higher of the two versions.
-        // This may not work! It's entirely possible that the result won't
-        // compile. We're going to try it and see what happens.
-        version = Version.getHigherVersion(version, existingVersion);
-      }
-      testElmJson["dependencies"]["direct"][name] = version;
-    });
-  }
-
-  function addIndirectDependencies(deps) {
-    Object.keys(deps).forEach(function(name) {
-      if (testElmJson["dependencies"]["direct"].hasOwnProperty(name)) {
-        // already a normal dep
-      } else {
-        var version = deps[name];
-
-        if (testElmJson["dependencies"]["indirect"].hasOwnProperty(name)) {
-          var existingVersion = testElmJson["dependencies"]["indirect"][name];
-
-          // If we have a clash, choose the higher of the two versions.
-          // This may not work! It's entirely possible that the result won't
-          // compile. We're going to try it and see what happens.
-          version = Version.getHigherVersion(version, existingVersion);
-        }
-
-        testElmJson["dependencies"]["indirect"][name] = version;
-      }
-    });
-  }
-
-  var nodeTestRunnerElmJsonPath = path.resolve(path.join(__dirname, "..", "elm.json"));
-  var nodeTestRunnerElmJson = fs.readJsonSync(nodeTestRunnerElmJsonPath);
-  // we want to use the version of elm-explorations/test that the user
-  // specifies in their own test-dependencies. everything else needs to
-  // be included for the test runner to compile.
-  delete nodeTestRunnerElmJson.dependencies.direct['elm-explorations/test'];
-  addDirectDependencies(nodeTestRunnerElmJson["dependencies"]["direct"]);
-  addIndirectDependencies(nodeTestRunnerElmJson["dependencies"]["indirect"]);
-
-  if (isPackageProject) {
-    addDirectDependencies(projectElmJson["dependencies"]);
-    addDirectDependencies(projectElmJson["test-dependencies"]);
-    // package projects don't explicitly list their transitive dependencies,
-    // to we have to figure out what they are.  We write the elm.json that
-    // we have so far, and run elm to see what it thinks is missing.
-    fs.writeFileSync(
-      path.join(generatedCodeDir, "elm.json"),
-      JSON.stringify(testElmJson, null, 4)
-    );
-    var missingDeps = askElmForMissingTransitiveDependencies(generatedCodeDir);
-    addIndirectDependencies(missingDeps);
-  } else {
-    addDirectDependencies(projectElmJson["dependencies"]["direct"]);
-    addIndirectDependencies(projectElmJson["dependencies"]["indirect"]);
-    addDirectDependencies(projectElmJson["test-dependencies"]["direct"]);
-    addIndirectDependencies(projectElmJson["test-dependencies"]["indirect"]);
-  }
-
-  // Make all the source-directories absolute, and introduce a new one.
-  var projectSourceDirs;
-  if (isPackageProject) {
-    projectSourceDirs = [ "./src" ];
-  } else {
-    projectSourceDirs = projectElmJson["source-directories"];
-  }
-  var sourceDirs = projectSourceDirs.map(
-    function(src) {
-      return path.resolve(path.join(projectRootDir, src));
-    }
-  ).concat(testRootDir);
-
-  testElmJson["source-directories"] = [
-    // Include elm-stuff/generated-sources - since we'll be generating sources in there.
-    generatedSrc,
-
-    // NOTE: we must include node-test-runner's Elm source as a source-directory
-    // instead of adding it as a dependency so that it can include port modules
-    path.resolve(path.join(__dirname, "..", "src"))
-  ].concat(sourceDirs);
-
-  // When running node-test-runner's own test suite, the node-test-runner/src folder
-  // will get added twice: once because it's the source-directory of the packge being tested,
-  // and once because elm-test will always add it.
-  // To prevent elm from being confused, we need to remove the duplicate when this happens.
-  testElmJson["source-directories"] =
-    testElmJson["source-directories"].filter(function(value, index, self) {
-      return self.indexOf(value) === index;
-    });
-
-  // Generate the new elm.json
-  fs.writeFileSync(
-    path.join(generatedCodeDir, "elm.json"),
-    JSON.stringify(testElmJson, null, 4)
-  );
-
-  return [generatedSrc, sourceDirs];
-}
-
-function askElmForMissingTransitiveDependencies(pathtoElmProject) {
-  var result = spawn.sync(pathToElmBinary, ["make", "--report=json"], {
-    silent: true,
-    cwd: pathtoElmProject
-  });
-
-  // TODO: hopefully the next Elm 0.19 beta will print the JSON to stdout instead
-  var output = result.stderr.toString();
-
-  // TODO: hopefully the next Elm 0.19 beta will not have this message (only occurs on linux) mixed in with the JSON output
-  output = output.replace(/^elm:.*no version information available \(required by elm\)\n/, '');
-
-  var report = JSON.parse(output);
-  if (report.type === "error" && report.title === "NO INPUT") {
-    // all transtive dependencies are listed already (elm is complaining that we didn't tell it which .elm file to build, which happens after the dependency validations)
-    return {};
-  } else if (report.type === "error" && report.title === "MISSING DEPENDENCIES") {
-    var missingDeps = {};
-    // parse the missing dependencies and versions from the error report
-    report.message[1].string.replace(/"([^"]*)": "([^"]*)"/g, function(_, name, version) {
-      missingDeps[name] = version;
-    })
-    return missingDeps;
-  } else {
-    console.error("elm-test internal error: got an unexpected result from 'elm make' when validating transitive dependencies.  Please report this at https://github.com/rtfeldman/node-test-runner/issues");
-    process.exit(1);
-    return {}; // This makes flow happy because it doesn't know process.exit will stop everything
-  }
-}
-
-function generateAndRunTests(tests, filePathArgs, generatedSrc, getGlobs) {
+function generateAndRunTests(tests, filePathArgs, generatedCodeDir, generatedSrc, getGlobs) {
   // Building things like:
   //
   // import MyTests
@@ -636,7 +454,7 @@ function generateAndRunTests(tests, filePathArgs, generatedSrc, getGlobs) {
   // if some of Main's dependencies (such as an individual test file) changed.
   fs.writeFileSync(mainFile, testFileContents);
 
-  return runTests(mainFile);
+  return runTests(generatedCodeDir, mainFile);
 }
 
 function getWatchPaths(projectRootDir) {
